Changed returning int from getBallType methods to returning an enum value
